(Post && pictures) fixed details

+++ assets/js/posts.js
@@ -2,27 +2,35 @@ let Posts = {
     box: document.getElementById("postBox"),
     area: document.getElementById("newPost"),
     share: document.getElementById("share"),
+    userIndex: localStorage.getItem("userIndex"),
     makeElement: function(elementName, styleClass, content){
         let element = document.createElement(elementName);
         element.classList.add(styleClass);
         element.innerHTML = content;
         return element;
     },
-    strucure: function(time, txt){
+    makeImg: function(imgSrc){
+        let element = document.createElement("img");
+        element.classList.add("post__img");
+        element.src = imgSrc;
+        return element;
+    },
+    strucure: function(time, txt, userMail, profilePicture, author){
         let post = Posts.makeElement("div", "post", "");
-        let img = Posts.makeElement("img", "post__img", "");
-        img.src = Users.storage[Pictures.userIndex].profilePicture[Users.storage[Pictures.userIndex].profilePicture.length - 1];
-        post.appendChild(img);
-        post.appendChild(Posts.makeElement("p", "post__author", Users.storage[Pictures.userIndex].name + " " + Users.storage[Pictures.userIndex].lastName));
+        let hiddenMail = Posts.makeElement("p", "post__author", userMail);
+        hiddenMail.style.display = "none";
+        post.appendChild(hiddenMail);
+        post.appendChild(Posts.makeImg(profilePicture));
+        post.appendChild(Posts.makeElement("p", "post__author", author));
         post.appendChild(Posts.makeElement("p", "post__time", time));
         post.appendChild(Posts.makeElement("p", "post__content", txt));
         return post;
     },
     draw: function(){
-        if(Users.storage[Pictures.userIndex].posts.length !== 0){
+        if(Statuses.storage.length !== 0){
             Posts.box.innerHTML = "";
-            for(let i = 0; i < Users.storage[Pictures.userIndex].posts.length; i++){
-                Posts.box.appendChild(Posts.strucure(Users.storage[Pictures.userIndex].posts[i].date, Users.storage[Pictures.userIndex].posts[i].content));
+            for(let i = 0; i < Statuses.storage.length; i++){
+                Posts.box.appendChild(Posts.strucure(Statuses.storage[i].date, Statuses.storage[i].content, Statuses.storage[i].eMail, Statuses.storage[i].profilePicture, Statuses.storage[i].author));
             }
         }
     },
@@ -49,36 +57,49 @@ let Posts = {
         return renderedDate;
     },
     deletePost: function(){
-        if(Posts.box.childElementCount == 1){
-            document.getElementsByClassName("post")[0].addEventListener("click", function(){
-                Users.storage[Pictures.userIndex].posts = [];
-                window.location.replace("logined.html");
-            });
-        }
-        for(let i = 0; i < Posts.box.childElementCount; i++){
-            let element = document.getElementsByClassName("post")[i];
-            element.addEventListener("click", function(){
-                let answer = confirm("Are you shure you want delete this post?");
-                if(answer){
-                    Users.storage[Pictures.userIndex].posts.splice(i, 1);
-                    Users.resetStorage();
-                    Posts.draw();
-                    Posts.deletePost();
+        if(Statuses.storage.length == 1){
+            let zeroIndexPost = Posts.box.childNodes[0];
+            if(zeroIndexPost.childNodes[0].innerHTML == Users.storage[Posts.userIndex].eMail){
+                Posts.box.childNodes[0].addEventListener("click", function(){
+                    let answer = confirm("Are you shure you want delete this post?");
+                    if(answer){
+                        Statuses.storage = [];
+                        Statuses.resetStorage();
+                        window.location.replace("logined.html");
+                    }
+                });
+            }
+        }else{
+            for(let i = 0; i < Posts.box.childElementCount; i++){
+                let element = document.getElementsByClassName("post")[i];
+                if(element.childNodes[0].innerHTML == Users.storage[Posts.userIndex].eMail){
+                    element.addEventListener("click", function(){
+                        let answer = confirm("Are you shure you want delete this post?");
+                        if(answer){
+                            Statuses.storage.splice(i, 1);
+                            Statuses.resetStorage();
+                            Posts.draw();
+                            Posts.deletePost();
+                        }
+                    });
                 }
-            });
+            }
         }
     },
     new: function(){
         Posts.share.addEventListener("click", function(){
             if(Posts.area.value !== ""){
-                Users.storage[Pictures.userIndex].posts.push(
+                Statuses.storage.push(
                     {
+                        author: Users.storage[Posts.userIndex].name + " " + Users.storage[Posts.userIndex].lastName,
                         date: Posts.getPostDate(),
                         content: Posts.area.value,
+                        eMail: Users.storage[Posts.userIndex].eMail,
+                        profilePicture: Users.storage[Posts.userIndex].profilePicture[Users.storage[Posts.userIndex].profilePicture.length - 1],
                     }
                 );
+                Statuses.resetStorage();
                 document.getElementById("newPost").value = "";
-                Users.resetStorage();
                 Posts.draw();
                 Posts.deletePost();
             }
